requirements.txt
flask
redis
mcstatus
requests
Pillow
discord.py
flask-cors